backend/requirements.txt
fastapi
uvicorn
pymongo[srv]
python-dotenv
passlib[bcrypt]
python-jose
scikit-learn
pandas
numpy
joblib
certifi
